records: convert staticmethods in classmethods
